Polyloader for Python 2 (2.6 and 2.7) is GREEN.  Woot!
Unit tests included to ensure that Polyloader doesn't interfere with
standard Python import behavior, while still allowing for arbitrary
compilers to be associated with new file suffixes on the fly.

WOOT!

I wrote unit tests!

The syntax of the compilers has been change; I chose to trust Python's
source file loaders, so now we get a default stream in whatever format
the current version of Python is most comfortable with.

Made a note in my TODO about a Context Manager for making Polyloaders
"live" and then removing them automatically when a procedure goes out
of scope.

I have this weird sensation that this might actually work...